Stats funzia

+++ ProgettoOOP/src/main/java/univpm/esame/ProgettoOOP/util/stats/StatsInterface.java
@@ -8,6 +8,6 @@ import univpm.esame.ProgettoOOP.model.AbstractObject;
 
 
 public interface StatsInterface  {
-    public HashMap<String, Integer> StatsType(ArrayList<AbstractObject> file) throws TypeNotRecognisedException;
+    public HashMap<String, Long> StatsType(ArrayList<AbstractObject> file) throws TypeNotRecognisedException;
     public HashMap<String, ?> StatsSize(ArrayList<AbstractObject> files) throws TypeNotRecognisedException;
 }
